Add files via upload
The new classes. The layout files will be sent in the next pull request. Sorry to send many pull requests like that. I modified some functions from the folder I sent you on google. I included everything that was on the previous database classes and I added my functions on it.